Indicate whether the bookmark has been visited
This inserts a MEDIUM BLACK CIRCLE (\u26AB) if the bookmark has
never been visited, or a MEDIUM WHITE CIRCLE (\u26AA) if it has.

This provides an easy visual indication of whether you have seen
a particular item already.

The indicators will not update until the next feed refresh.